inmeta-service/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
httpx>=0.25.2
pandas>=2.1.3
redis>=5.0.1
celery>=5.3.6
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx-mock>=0.11.0
